backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
redis==6.4.0
arq==0.26.3
pgvector==0.4.1
pypdf==6.0.0
python-docx==1.2.0
email-validator==2.3.0
pytest==8.4.1
pytest-asyncio==1.1.0
aiosqlite==0.21.0
